algolibs: refactored tests, added shift functionality to maze graph

+++ algolibs/benchmark/benchmark.cpp
@@ -46,7 +46,7 @@ void runBenchmark(const MazeGraph & graph, size_t runs, QuerySupplier query_supp
 std::vector<std::pair<Location, Location>> createSnakeGraphQueries(size_t extent, size_t number)
 {
     std::default_random_engine rng;
-    std::uniform_int_distribution<int> dist(0, extent - 1);
+    std::uniform_int_distribution<int> dist(0, static_cast<int>(extent - 1));
     std::vector<std::pair<Location, Location>> queries;
     queries.reserve(number);
     for (auto i = 0; i < number / 2; i++) {
@@ -60,7 +60,7 @@ std::vector<std::pair<Location, Location>> createSnakeGraphQueries(size_t extent
 std::vector<std::pair<Location, Location>> createTreeGraphQueries(size_t extent, size_t number) {
     std::vector<std::pair<Location, Location>> queries;
     std::default_random_engine rng;
-    std::uniform_int_distribution<int> dist(0, extent - 2);
+    std::uniform_int_distribution<int> dist(0, static_cast<int>(extent - 2));
     queries.reserve(number);
     for (auto i = 0; i < number / 2; i++) {
         queries.push_back(std::make_pair(Location(extent - 1, dist(rng)), Location(dist(rng), 0)));
